Fix several routines in sphnfs.f90
There was a bug in the use of grid_selection and the subsequent logic to identify
the best grid. In the call of grid_selection spHNFs was used instead of best_HNF.
So best_HNF was never set but later used in the condition:
if (any(best_HNF > 0)) then
This lead the code run for eternity, when using intel with -O3, since
uninitialized variables are set to 0.

I think I fixed it for all occurrences.